CoRider_Flask/requirements.txt
Flask
python-dotenv
pymongo
Flask-RESTful
Flask-Bcrypt
dnspython
